A simple Java survey maker/taker
Asks one user to enter questions and the number of people that will be
taking the survey. Then the people take the survey. they are shown the
questions previously entered and are asked to rate their response in a
scale of 1-5. Then all of the respondents answers are shown in a matrix.
Then the user can review the answers for any particular question
individually.